+ move social algorithm to its own module
